daily-draw/pyproject.toml
[project]
name = "daily-draw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.0",
]
